requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
requests==2.31.0
python-multipart==0.0.6
Pillow==10.2.0
